Turn-Base Updates
Added some changes to the script, still trying to assign it to the black/purple blob.
Added a Hero script so that the turn base script searches for the "Hero" character
Look at it later to see it, you can make any changes you see fit, I plan on working more on it tomorrow.