Combine Script
The script is done, it works both on overworld and dungeon. As in spawning the player
Overworld --> Movement works but I can phase through the map now not sure why. I checked the order in layer + the sorting layer and still can phase through the map

Dungeon --> I spawn but the other scripts don't recognize the player something about the CombatantController but thats a script you two made and wont edit or change anything unless one or both of you are there

Those are the only two issue, I'm facing right now. Aside from that the scripts copies the entire components from both pre-fabs and spawns them in once the game loads.
Colliders work + the dungeon teleport works when you press E