Enemy Spawning Functional
Enemy count ranges functional.
- EnemyPool class
- EnemyPoolElement class
- Can assign different max numbers to each EnemyPoolElement
- EnemyPoolElement has a checkbox for IsDefault
- EnemyPool fills in any remaining sapces in its list of enemies to spawn with copies of the first EnemyPoolElement which has IsDefault checked.

This functionality has been tested with all 3 difficulties using 3 presets.

Made copies of the Goblin placeholder enemy to create different colors for testing.

GAME.MGR is checked to preserve through scenes (only changed this in the neighborhood scene)

New component GoToNeighborhood is a standalone script with a single function that calls GoToNeighborhood on whatever is the current Singleton instance of GAME
- TODO : Do above for GoToDungeon.
- This way anything (keypress, button, etc) can call this function at runtime programmatically instead of needing to use UnityEvents to find a specific instance of a Singleton (which defeats the purpose)