Shield Bash
Sending in this Shield Bash still needs to be tested. I'm switching over to my laptop to finish it and finish the code for regeneration.
Added folders for each class so we can put their abilities in there respective folders.
Once again, I'll finish the regen on my laptop by tonight.