Really bad solution to a bug
Interaction prompts rely on the interacted object to define a string that will be displayed when the interaction prompt pops up.  Apparently, no dungeon entrances have defined this.
Here, I've just written in a few lines to check the parent for a `DungeonEntrance` if their `promptString == null || promptString == ""`.  If it finds a dungeon entrance, it displays "to Enter".

lmao