Added a logger
`class dbug` is a serializable class that
- Wraps up standard Unity `Debug.Log[etc].(string s)` functions
- Exposed bool in the editor for on or off.
- Easily turn console debugging on and off per object.

```
[SerializeField] dbug log;

public void SomeMethod(MyClass someObj)
{
    log.print("hi");
    log.warn("im warning u");
    log.error("oh this is bad");

    // do stuff
}
```

Under "Log" (or whatever you wanna call it) in the inspector of whatever you put the `dbug` on, youll see a bool `showMessages`.  If that is checked, you'll see everything above printed to the console.  If not, you'll just `// do stuff`