Save Script
Made the script, had to change currency to credits and all variations of it.
Player Manager has the instance for the save script although, I need to find the location of each save feature and instance it so the code I'm sending is blocked out but its there for you guys to review.