Broadened the tool.
Can now redirect any field to another field of a matching type, hold the reference, and then safely delete the fields.  Finds all prefabs in the project that implement a specific mono script and performs this on all of them.
Should still probably only be used only on separate branches until we're more confident that it's not like.. volatile.

I will test it out on dungeons tomorrow.