Fix Load Screen
Changed UI.MGR.OpenLoadScreen() to be called from
IntersectionManager.MGR.StartGenerationAt() (or whatever the fn is
called that starts the actual generation of the environ. The function
gets called on Start() and on Regenerate(), so it should cover our
bases).
UI.MGR still subscribes to IntersectionManager.MGR.GenerationComplete
independent of the call to OpenLoadScreen(), and
UI.MGR.HandleGenerationComplete() includes
  - Unsubscribing from IntersectionManager.MGR.GenerationComplete
  - Closing the LoadingScreen